+++ CHANGELOG.rst
@@ -1,4 +1,6 @@
 Master:
+    - Addef ``RecordList.prefetch`` method. Als *RecordList's* *fields* argument
+      now works.
     - Changed ``Object`` class. Now it have ``extend_me.ExtensibleByHashType`` metaclass
       which allows it to be extended separatly for each model, and in general way.
       For example of extending for specific model look at ``openerp_proxy.plugins.module_utils``